Fix float-to-Double compilation error in CosmosTotalResultReporter

+++ sdk/cosmos/azure-cosmos-benchmark/src/main/java/com/azure/cosmos/benchmark/CosmosTotalResultReporter.java
@@ -107,7 +107,7 @@ public class CosmosTotalResultReporter {
         double intervalInSeconds = Duration.between(lastRecorded, now).toMillis() / 1000.0;
 
         if (intervalInSeconds > 0) {
-            cpuUsageSamples.add(cpuReader.getSystemWideCpuUsage());
+            cpuUsageSamples.add((double) cpuReader.getSystemWideCpuUsage());
 
             if (successCount == 0 && failureCount == 0) {
                 lastRecorded = now;
